+++ radio.h
@@ -12,7 +12,7 @@ public:
 
     virtual void play(QString url) = 0;
     virtual void stop() = 0;
-    virtual void setVolume(double volume) = 0;
+    virtual void setVolume(int volume) = 0;
 
 signals:
     void bufferChanged(int value);
